Cric 1636 redesign submission tab layout (#1538)
* CRIC-1636 Redesign Submission Tab layout revert changes

* Custom width for submission tab

* adding custom width submission tab

* Added actions column in submissions tab

* Added actions column in submissions tab

* Added tooltip text in action column

* Added tooltip in actions column

* Imported react tooltip

* Removed tooltip component

* Added tooltip component in submissions tab

* Added tooltip component in submission tab

* Number column heading changed

* Added tooltip component in submission tab

* Added tooltip component in submission tab
